requirement.txt
asgiref==3.8.1
certifi==2025.1.31
charset-normalizer==3.4.1
crispy-tailwind==1.0.3
Django==5.1.5
django-allauth==65.3.1
django-crispy-forms==2.3
idna==3.10
pillow==11.1.0
psycopg2-binary==2.9.10
python-dotenv==1.0.1
requests==2.32.3
sqlparse==0.5.3
tzdata==2025.1
urllib3==2.3.0
